Allow users to specify no capacity
The new [Basic SKU][0] does not support any scaling, and it is an error to
provide `Capacity` or `AutoScaleSettings`. The resource is implemented
in such a way that it was not possible to call
`DeploymentsCreateOrUpdateThenPoll` without one of these in the
`ScalingProperties`, which is illegal for the Basic SKUs.

Currently the backend has a hack to silently drop `ScalingProperties`
for Basic SKUs, but this is a confusing user experience:

- initial `terraform apply` submits a `PUT` with `"scalingProperties":
{"capacity": 20}` - this validation error is currently suppressed and
the server reports back `"scalingProperties": null`

- next `terraform plan` detects a `capacity 0 -> 20`

- applying that plan requests a `PATCH` with `"scalingProperties":
{"capacity": 20}`, that fails with a `UnsupportedOnBasicPlan` error

It gets more misleading if users specify a capacity explicity;
`capacity = 500` or `auto_scale_profile` will apply successfully, but
get silently ignored server-side.

This patch loosens the `capacity`'s `ValidateFunc` to permit 0, letting
users specify no capacity. The existing `DeploymentResource` will omit
the `scalingProperties` on `PUT` and `PATCH` when it sees a zero.

This has a side benefit of making it obvious that basic skus are not
meant for production workloads.

Considered a few alternatives, but this one seemed best:

- removing the `Default: 20`: ran into other server-side bugs,
backward compatibility issues, and some thorny UX issues if someone
adds an explicit capacity, and then removes it

- adding special cases for basic skus (i.e. `if strings.HasPrefix(sku,
"basic")`: duplicating non-trivial server-side validation logic in the
TF client smell like maintenance headaches and puts an impractical
onus on users to update their `azurerm` provider as the validation
rules evolve

[0]: https://docs.nginx.com/nginxaas/azure/billing/overview/#basic-plan