nginx_deployment_resource: improve Basic SKU support
The new Basic SKU does not support any scaling, and it is an error to
provide `Capacity` or `AutoScaleSettings`. The resource is implemented
in such a way that it was not possible to call
`DeploymentsCreateOrUpdateThenPoll` without one of these in the
`ScalingProperties`, which is illegal for the Basic SKUs.

Currently the backend has a hack to silently drop `ScalingProperties`
for Basic SKUs, but this is a confusing user experience:

- Basic SKU users who don't specify `capacity` can successfully
`terraform apply`, but every `terraform plan` tries to update
`capacity 0 -> 20`. We tell users to `ignore_changes` to workaround
this.

- Basic SKU users who do specify `capacity` or `auto_scale_profile`
can successfully `terraform apply`, but those changes don't have any
effect on the resources in Azure. Every `terraform plan` will try to
add those scaling settings back.

Our `capacity` field had a default value of 20, which matches the
server-side defaults for the Standard plan, but that default is not
globally applicable anymore.

This patch removes the default value, and adjusts the diff and update
logic to maintain backwards compatibility for Standard plan users:

- if the server says 20 and the user has not requested any `capacity`,
no updates are needed

- if the user removes both `capacity` and `auto_scale_settings`, we
send an empty (but non-nil) `ScalingProperties` in the PATCH request,
which restores the server's default values for that resource's `sku`

As a happy side-effect, `auto_scale_settings` users no longer need to
use `ignore_changes` to mask off `capacity`; terraform and the server
models will agree.

Once this patch is released, NGINX will remove our `ScalingProperties`
hack, and users will get informative errors from our server-side
validation instead of misleading successes.